HMS/requirements.txt
Django==4.2.2
django_environ==0.10.0
django_filter==23.2
djangorestframework==3.14.0
